+++ expertise/preprocess/textrank_words.py
@@ -1,4 +1,4 @@
-from expertise.preprocessors.textrank import TextRank
+from expertise.preprocess.textrank import TextRank
 
 def keyphrases(text, include_scores=False, include_tokenlist=False):
 	textrank = TextRank()
